feat: add stories for atomic
