+++ .storybook/main.js
@@ -1,7 +1,5 @@
 module.exports = {
   stories: [
-    '../stories/**/*.stories.mdx',
-    '../stories/**/*.stories.@(js|jsx|ts|tsx)',
     '../packages/atomic/src/**/*.stories.@(js|jsx|ts|tsx)',
     '../packages/uikits/src/**/*.stories.@(js|jsx|ts|tsx)',
     '../packages/layout/src/**/*.stories.@(js|jsx|ts|tsx)',
