--- packages/atomic/__test__/a.test.mjs
@@ -1,6 +1,0 @@
-import { styled, getCssText } from '../src/stitches';
-import { Box } from '../src';
-
-const x = styled(Box, {});
-console.log(getCssText());
-console.log(x);
