+++ packages/atomic/src/Link/index.tsx
@@ -1,24 +1,2 @@
-import { Atomic } from '../Atomic';
-import { link } from './link.utils';
-import React from 'react';
-import { BaseProps, styled } from '@unstyled-ui/core';
-import classNames from 'classnames';
-
-export type LinkProps = BaseProps<React.LinkHTMLAttributes<HTMLLinkElement>> & {
-  to: string;
-};
-
-export const Link: React.FC<LinkProps> = (props: LinkProps) => {
-  const { css, className, to, ...restProps } = props;
-  return (
-    //@ts-ignore
-    <Atomic
-      as="a"
-      className={classNames(className, 'c3-link')}
-      //@ts-ignore
-      css={{ ...link(), ...css }}
-      href={to}
-      {...restProps}
-    />
-  );
-};
+export * from './Link';
+export * from './link.utils';
