+++ packages/atomic/src/Text/Text.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Text } from '../..';
+import { Text } from '..';
 const Template = (args: any) => <Text {...args} />;
 
 export const TwoLineText = Template.bind({}) as any as { args: any };
